Work in progress - adding dynamic persistence
