+++ src/main/java/persistence/CalculationPersistence.java
@@ -1,6 +1,7 @@
 package persistence;
 
 import model.CalculationModel;
+import model.ConcitoItemModel;
 import org.hibernate.Session;
 
 public class CalculationPersistence {
@@ -9,6 +10,7 @@ public class CalculationPersistence {
 
     public static void create(){
         //TODO
+
 
     }
 
@@ -21,6 +23,10 @@ public class CalculationPersistence {
         read.beginTransaction();
         read.getTransaction().commit();
         CalculationModel gaming = read.find(CalculationModel.class, 1);
+        Integer id = (Integer) read.save(gaming);
+        ConcitoItemModel hej = new ConcitoItemModel("XD", 32.0, "XDD", "XDDD");
+        Integer id2 = (Integer) read.save(hej);
+        System.out.println("HER ER ID: " + id + id2);
         System.out.println(gaming.calcAveCO2prKg());
         System.out.println(gaming.calcTotalCo2());
         System.out.println(gaming.calcTotalKg());
