--- src/main/java/persistence/FoodDescriptionPersistence.java
@@ -1,25 +1,0 @@
-package persistence;
-
-public class FoodDescriptionPersistence {
-
-//Work in progress
-
-    public static void create(){
-        //TODO
-    }
-
-    public static void update(){
-        //TODO
-    }
-
-    public static void read(){
-        //TODO
-    }
-
-    public static void delete(){
-        //TODO
-    }
-
-
-
-}
